requirements.txt
groq
python-dotenv
customtkinter
networkx
sentence-transformers